Fix translation issues
Some things were correctly translating, but most things were not.
Especially on the admin panel.

The issue seemed to be loading the load_plugin_textdomain in the
init hook in the file init.php doesn't work correctly.
Moved it to the file variables.php in the plugins_loaded hook.
The translations now load at the correct time in the
initialization sequence and tests are successful.