Update .pot language template
The .pot file was out of date and missing quite a few i18n lines.

I regenerated the language template and synced the Author, License,
Description, etc. from the original. About 160 new translations,
a handful of case changes and a few removals that are no longer
present in the source code.