feat: Update database tables with approximate location boolean
Migrates database to support tracking approximate locations. This will run once
when the admin updates to this version (probably 3.9.0) and begins tracking `tsml_version`
in the `wp_options` table.

Updates `tsml_addresses` (in `wp_options`) cache and appropriate location meta data with
`is_approximate_location` when a `formatted_address` is deemed to be approximate. This is done
to enable the correct settings for the location and to display the map without directions, a marker,
or info boxes.

see #173