fix: Hot-fix to reduce invalid geocoding calls (errors)
The `tsml_get_meeting` was called each time a user started
a new meeting even when address values were empty. Because there were no
checks to see if fields were empty, Google was called with an empty address.

A search was made to see if other blocks of code were unexpectedly calling
`tsml_get_meeting`, but none were found.

This commit should help with the issue documented in #255, but only to the
extend that the extra geocodes caused when a meeting is manually being
entered are avoided.