Fix for #257 - Choosing Sunday shows no meetings found.
The specific issue here is that 11th step meetings store the value "11" as an int in the "types" array instead of a string.  This is passed to sanitizeTitle() which tries to apply a regex to the value, and this produces a type error since the value is an int rather than a string.

To protect against this specific case and also any other meeting info fields that might have the issue, added conversion to string of the passed value to sanitizeTitle() and sanitizeDataSort().