feat: Add attendance options to query string
To make use of the new attendance options for taylored meeting list pages,
this feature add support to include `&tsml-attendance_option=` in the
query string. The code checks to ensure the option is one of the four
choices set in `tsml_meeting_attendance_options`.

TODO: make the query string persistent once the jQuery calls are invoked.

see #312