refactor: Sanitize attendance options earlier/improve filtering
Previous approach checked and sanitized the attendance options provided
in the URL when filtering results. Now we sanitize them when parsing
the initial query. This is safer and more consistent with how types are
parsed.

This also better handles multipe attendance options, and `filter_meetings`
now adds `hybrid` automatically if either `online` or `in_person` are
requested.

see #312