fix: Implement URL parsing for 'active' meetings.
v3.11.0 did not parse the URL query for attendance_option='active'
because this is not a pure attendance option. Rather, it is a superset
of online, in_person, hybrid types.

This fix enables admins to properly create a meeting list of only active
meetings (i.e., not inactive), and this choice persists into the user dropdown
for meeting types.

closes #396